ci: harden golangci-lint and clear surfaced issues
Add gocritic, gocyclo, gocognit, errorlint, nakedret, nolintlint,
prealloc, unparam, whitespace, usestdlibvars, gosec, plus the
gofumpt formatter to the lint config. Exempt test files from the
noisier rules. Fix the issues this surfaced: gocritic truncateCmp
in pysync (compare uint32 values instead of narrowing to uint8),
gocritic unnamedResult on captureFile, and gofumpt formatting in
hash/secret_test.go.

+++ pysync/mutex.go
@@ -129,9 +129,9 @@ func (m *Mutex) LockTimed(timeout time.Duration, flags LockFlags) LockStatus {
 			}
 		}
 
-		expected := newv
+		expected := uint32(newv)
 		ret := Park(unsafe.Pointer(&m.bits), func() bool {
-			return uint8(m.bits.Load()) == expected
+			return m.bits.Load() == expected
 		}, timeout, entry, flags&LockDetach != 0)
 
 		switch ret {
